Add files via upload
https://contest.yandex.ru/contest/68397/run-report/120178692/
Implement a naive binary search tree for integers. The program receives a sequence of integers as input and builds a tree from them. Elements are added to the tree sequentially according to their input. If the element already exists in the tree, you do not need to add it. Tree balancing is not performed.

For the resulting tree, output a list of all leaves (vertices that have no descendants) in ascending order.